+++ lib/project/pro_motion/fragments/refreshable_list.rb
@@ -1,34 +1,39 @@
-# # Should be included in PMListScreen but RMA is acting up
 module RefreshableList
 
-#   def make_refreshable(params={})
-#     @ptr = find!(In::Srain::Cube::Views::Ptr::PtrFrameLayout)
-#     header = In::Srain::Cube::Views::Ptr::Header::MaterialHeader.new(app.context)
-#     ptr_colors = params[:color_array] || [color.dark_gray]
-#     header.setColorSchemeColors(ptr_colors)
-#     header.setPtrFrameLayout(@ptr)
-#     @ptr.setHeaderView(header)
-#     @ptr.addPtrUIHandler(header)
-#     @ptr.ptrHandler = PTRHandler.new do |frame|
-#       on_refresh if respond_to?(:on_refresh)
-#     end
-#   end
+  def self.included(base)
+    base.extend(ClassMethods)
+  end
 
-#   def self.refreshable(params = {})
-#     @refreshable_params = params
-#     @refreshable = true
-#   end
+  module ClassMethods
+    def refreshable(params = {})
+      @refreshable_params = params
+      @refreshable = true
+    end
 
-#   def self.get_refreshable
-#     @refreshable ||= false
-#   end
+    def get_refreshable
+      @refreshable ||= false
+    end
 
-#   def self.get_refreshable_params
-#     @refreshable_params ||= nil
-#   end
+    def get_refreshable_params
+      @refreshable_params ||= nil
+    end
+  end
 
-#   def stop_refreshing
-#     @ptr.refreshComplete
-#   end
+  def make_refreshable(params={})
+    @ptr = find!(In::Srain::Cube::Views::Ptr::PtrFrameLayout)
+    header = In::Srain::Cube::Views::Ptr::Header::MaterialHeader.new(app.context)
+    ptr_colors = params[:color_array] || [color.dark_gray]
+    header.setColorSchemeColors(ptr_colors)
+    header.setPtrFrameLayout(@ptr)
+    @ptr.setHeaderView(header)
+    @ptr.addPtrUIHandler(header)
+    @ptr.ptrHandler = PTRHandler.new do |frame|
+      on_refresh if respond_to?(:on_refresh)
+    end
+  end
+
+  def stop_refreshing
+    @ptr.refreshComplete
+  end
 
 end
